Add one book and one blog post about strategy
The book Good Strategy/Bad strategy uses fascinating examples from business, nonprofit, and military affairs to bring its original and pragmatic ideas to life. It's one of the best books about strategy.

The post is based on the book above, but more oriented to software engineering. The author is @rpfelgueiras